+++ changelog.md
@@ -1,3 +1,11 @@
+1.18.2
+------
+- add missing options from dm_event.fields enum #547 (@RenKoya1)
+
+1.18.1
+------
+- feat: add most_recent_tweet_id field #545 (@RenKoya1)
+
 1.18.0
 ------
 - feat: update again domain from twitter.com to x.com #532 (@melvinmcrn)